requirements.txt
Flask==2.0.1
PyNaCl==1.4.0
requests==2.25.1
python-dotenv==0.18.0